requirements.txt
astropy==5.1
numpy==1.22.4
Pillow==9.3.0
reproject==0.8
scikit_image==0.19.3
skimage==0.0
